Removed egregious text from tutorial
No offense; but you really need to think twice if the text is necessary.
There's other text with the "can't dash through!" hazard saying "some walls can't be dashed through" ... Isn't that most walls? I think the player has dealt with this before-- and why is it a hazard?! It's a hazard that kills you?! We have that! It's called a spike! It's just a spike reskin! WHY IS THERE A TUTORIAL FOR THE SPIKE SIMPLY BECAUSE IT LOOKS DIFFERENT